Use HTML-first indentation
Tumblr themes are HTML first and theme logic second. Thus, they should be written in a way where it's easier to read the HTML than the theme logic, instead of the other way around. It may look like a hot fucking mess like this now, but that's entirely the point!